Fixed _filelengthi64 linux implmentation.  Also added ifdef around common.h
